refactor: 🔄 views.ts => views/[slug].ts
